+++ src/mpd/commands/current_song.rs
@@ -1,5 +1,7 @@
 use std::{collections::HashMap, time::Duration};
 
+use anyhow::Context;
+use chrono::{DateTime, Utc};
 use serde::Serialize;
 
 use super::metadata_tag::MetadataTag;
@@ -12,6 +14,9 @@ pub struct Song {
     pub duration: Option<Duration>,
     pub metadata: HashMap<String, MetadataTag>,
     pub stickers: Option<HashMap<String, String>>,
+    pub last_modified: DateTime<Utc>,
+    // Option because it is present from mpd 0.24 onwards
+    pub added: Option<DateTime<Utc>>,
 }
 
 impl std::fmt::Debug for Song {
@@ -37,6 +42,14 @@ impl FromMpd for Song {
                 self.duration = Some(Duration::from_secs_f64(value.parse().logerr(key, &value)?));
             }
             "time" | "format" => {} // deprecated or ignored
+            "last-modified" => {
+                self.last_modified =
+                    value.parse().context("Failed to parse date").logerr(key, &value)?;
+            }
+            "added" => {
+                self.added =
+                    Some(value.parse().context("Failed to parse date").logerr(key, &value)?);
+            }
             key => {
                 self.metadata
                     .entry(key.to_owned())