+++ src/ui/panes/mod.rs
@@ -280,8 +280,6 @@ pub(crate) trait Pane {
     }
 }
 
-pub mod dirstack {}
-
 pub(crate) mod browser {
     use std::{borrow::Cow, cmp::Ordering};
 
@@ -290,9 +288,10 @@ pub(crate) mod browser {
         style::{Color, Style},
         text::{Line, Span},
     };
+    use unicase::UniCase;
 
     use crate::{
-        config::theme::properties::SongProperty,
+        config::sort_mode::{SortMode, SortOptions},
         mpd::commands::{Song, lsinfo::LsInfoEntry},
         shared::mpd_query::PreviewGroup,
     };
@@ -378,6 +377,28 @@ pub(crate) mod browser {
                 );
             }
 
+            info_group.push(
+                Line::from(vec![
+                    start_of_line_spacer.clone(),
+                    Span::styled("Last Modified", key_style),
+                    separator.clone(),
+                    Span::from(self.last_modified.to_string()),
+                ])
+                .into(),
+            );
+
+            if let Some(added) = &self.added {
+                info_group.push(
+                    Line::from(vec![
+                        start_of_line_spacer.clone(),
+                        Span::styled("Added", key_style),
+                        separator.clone(),
+                        Span::from(added.to_string()),
+                    ])
+                    .into(),
+                );
+            }
+
             let mut tags_group = PreviewGroup::new(Some(" --- [Tags]"));
             for (k, v) in self
                 .metadata
@@ -406,96 +427,157 @@ pub(crate) mod browser {
 
     #[derive(Debug, Clone, PartialEq, Eq)]
     pub(crate) enum DirOrSong {
-        Dir { name: String, full_path: String },
+        Dir { name: String, full_path: String, last_modified: chrono::DateTime<chrono::Utc> },
         Song(Song),
     }
 
     impl DirOrSong {
         pub fn name_only(name: String) -> Self {
-            DirOrSong::Dir { name, full_path: String::new() }
+            DirOrSong::Dir { name, full_path: String::new(), last_modified: chrono::Utc::now() }
         }
 
         pub fn dir_name_or_file_name(&self) -> Cow<str> {
             match self {
-                DirOrSong::Dir { name, full_path: _ } => Cow::Borrowed(name),
+                DirOrSong::Dir { name, .. } => Cow::Borrowed(name),
                 DirOrSong::Song(song) => Cow::Borrowed(&song.file),
             }
         }
-    }
 
-    impl Ord for DirOrSong {
-        fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-            match (self, other) {
-                (DirOrSong::Dir { name: a, .. }, DirOrSong::Dir { name: b, .. }) => a.cmp(b),
-                (DirOrSong::Song(_), DirOrSong::Dir { .. }) => Ordering::Greater,
-                (DirOrSong::Dir { .. }, DirOrSong::Song(_)) => Ordering::Less,
-                (DirOrSong::Song(a), DirOrSong::Song(b)) => a.cmp_by_prop(b, &SongProperty::Title),
+        pub fn last_modified(&self) -> chrono::DateTime<chrono::Utc> {
+            match self {
+                DirOrSong::Dir { last_modified, .. } => *last_modified,
+                DirOrSong::Song(song) => song.last_modified,
             }
         }
     }
 
-    impl PartialOrd for DirOrSong {
-        fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-            Some(self.cmp(other))
-        }
-    }
-
     #[derive(Debug, PartialEq, Eq)]
-    pub(crate) struct SongCustomSort<'a>(&'a Song, &'a [SongProperty]);
+    pub(crate) struct SongCustomSort<'a, 'opts> {
+        song: &'a Song,
+        opts: &'opts SortOptions,
+    }
 
     impl Song {
-        pub(crate) fn with_custom_sort<'a>(
-            &'a self,
-            sort_props: &'a [SongProperty],
-        ) -> SongCustomSort<'a> {
-            SongCustomSort(self, sort_props)
+        pub(crate) fn with_custom_sort<'song, 'opts>(
+            &'song self,
+            opts: &'opts SortOptions,
+        ) -> SongCustomSort<'song, 'opts> {
+            SongCustomSort { song: self, opts }
         }
     }
 
-    impl Ord for SongCustomSort<'_> {
+    impl Ord for SongCustomSort<'_, '_> {
         fn cmp(&self, other: &Self) -> Ordering {
-            for prop in self.1 {
-                let ord = self.0.cmp_by_prop(other.0, prop);
-                if ord != Ordering::Equal {
-                    return ord;
+            match &self.opts.mode {
+                SortMode::Format(items) => {
+                    for prop in items {
+                        let ord = self.song.cmp_by_prop(other.song, prop);
+                        if ord != Ordering::Equal {
+                            return if self.opts.reverse { ord.reverse() } else { ord };
+                        }
+                    }
+                    Ordering::Equal
+                }
+                SortMode::ModifiedTime => {
+                    let result = self.song.last_modified.cmp(&other.song.last_modified);
+                    if self.opts.reverse { result.reverse() } else { result }
                 }
             }
-            Ordering::Equal
         }
     }
 
-    impl PartialOrd for SongCustomSort<'_> {
+    impl PartialOrd for SongCustomSort<'_, '_> {
         fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
             Some(self.cmp(other))
         }
     }
 
     #[derive(Debug, PartialEq, Eq)]
-    pub(crate) struct DirOrSongCustomSort<'a>(&'a DirOrSong, &'a [SongProperty]);
+    pub(crate) struct DirOrSongCustomSort<'dirsong, 'opts> {
+        dir_or_song: &'dirsong DirOrSong,
+        opts: &'opts SortOptions,
+    }
 
     impl DirOrSong {
-        pub(crate) fn with_custom_sort<'a>(
-            &'a self,
-            sort_props: &'a [SongProperty],
-        ) -> DirOrSongCustomSort<'a> {
-            DirOrSongCustomSort(self, sort_props)
+        pub(crate) fn with_custom_sort<'dirsong, 'opts>(
+            &'dirsong self,
+            opts: &'opts SortOptions,
+        ) -> DirOrSongCustomSort<'dirsong, 'opts> {
+            DirOrSongCustomSort { dir_or_song: self, opts }
         }
     }
 
-    impl Ord for DirOrSongCustomSort<'_> {
+    impl Ord for DirOrSongCustomSort<'_, '_> {
         fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-            match (self.0, other.0) {
-                (DirOrSong::Dir { name: a, .. }, DirOrSong::Dir { name: b, .. }) => a.cmp(b),
-                (DirOrSong::Song(_), DirOrSong::Dir { .. }) => Ordering::Greater,
-                (DirOrSong::Dir { .. }, DirOrSong::Song(_)) => Ordering::Less,
-                (DirOrSong::Song(a), DirOrSong::Song(b)) => {
-                    a.with_custom_sort(self.1).cmp(&b.with_custom_sort(self.1))
+            // dirs go first if group_directories_first is true
+            if self.opts.group_directories_first {
+                let type_order = match (self.dir_or_song, other.dir_or_song) {
+                    (DirOrSong::Song(_), DirOrSong::Dir { .. }) => Some(Ordering::Greater),
+                    (DirOrSong::Dir { .. }, DirOrSong::Song(_)) => Some(Ordering::Less),
+                    _ => None,
+                };
+
+                if let Some(order) = type_order {
+                    return if self.opts.reverse { order.reverse() } else { order };
                 }
             }
+
+            let order = match &self.opts.mode {
+                // directory grouping was already accounted above so we can now naively sort by
+                // mtime
+                SortMode::ModifiedTime => {
+                    self.dir_or_song.last_modified().cmp(&other.dir_or_song.last_modified())
+                }
+                // compare dirs against dirs, dirs against songs and songs against songs
+                SortMode::Format(items) => match (self.dir_or_song, other.dir_or_song) {
+                    (DirOrSong::Dir { name: a, .. }, DirOrSong::Dir { name: b, .. }) => {
+                        UniCase::new(a).cmp(&UniCase::new(b))
+                    }
+                    (DirOrSong::Song(a), DirOrSong::Song(b)) => {
+                        let ord = a.with_custom_sort(self.opts).cmp(&b.with_custom_sort(self.opts));
+                        // have to undo the reverse here because songs' custom sort reverses already
+                        if self.opts.reverse { ord.reverse() } else { ord }
+                    }
+                    (DirOrSong::Song(song), DirOrSong::Dir { name, .. }) => {
+                        for prop in items {
+                            let s = song.format(prop, "");
+                            if let Some(s) = s {
+                                let result = UniCase::new(s.as_ref()).cmp(&UniCase::new(name));
+                                if result != Ordering::Equal {
+                                    return if self.opts.reverse {
+                                        result.reverse()
+                                    } else {
+                                        result
+                                    };
+                                }
+                            }
+                        }
+                        Ordering::Greater
+                    }
+                    (DirOrSong::Dir { name, .. }, DirOrSong::Song(song)) => {
+                        for prop in items {
+                            let s = song.format(prop, "");
+                            if let Some(s) = s {
+                                let result =
+                                    UniCase::new(name.as_str()).cmp(&UniCase::new(s.as_ref()));
+                                if result != Ordering::Equal {
+                                    return if self.opts.reverse {
+                                        result.reverse()
+                                    } else {
+                                        result
+                                    };
+                                }
+                            }
+                        }
+                        Ordering::Less
+                    }
+                },
+            };
+            return if self.opts.reverse { order.reverse() } else { order };
         }
     }
 
-    impl PartialOrd for DirOrSongCustomSort<'_> {
+    impl PartialOrd for DirOrSongCustomSort<'_, '_> {
         fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
             Some(self.cmp(other))
         }
@@ -504,41 +586,334 @@ pub(crate) mod browser {
     impl From<LsInfoEntry> for Option<DirOrSong> {
         fn from(value: LsInfoEntry) -> Self {
             match value {
-                LsInfoEntry::Dir(crate::mpd::commands::lsinfo::Dir { path, full_path, .. }) => {
-                    Some(DirOrSong::Dir { name: path, full_path })
-                }
+                LsInfoEntry::Dir(crate::mpd::commands::lsinfo::Dir {
+                    path,
+                    full_path,
+                    last_modified,
+                }) => Some(DirOrSong::Dir { name: path, full_path, last_modified }),
                 LsInfoEntry::File(song) => Some(DirOrSong::Song(song)),
                 LsInfoEntry::Playlist(_) => None,
             }
         }
     }
+}
 
-    #[cfg(test)]
-    mod test {
+#[cfg(test)]
+#[allow(clippy::unwrap_used)]
+mod ordtest {
+    use std::{
+        sync::{LazyLock, atomic::AtomicU32},
+        time::Duration,
+    };
 
-        use super::DirOrSong;
-        use crate::mpd::commands::Song;
+    use itertools::Itertools;
 
-        #[test]
-        fn dir_before_song() {
-            let mut input = vec![
-                DirOrSong::Song(Song::default()),
-                DirOrSong::Dir { name: "a".to_owned(), full_path: String::new() },
-                DirOrSong::Song(Song::default()),
-                DirOrSong::Dir { name: "z".to_owned(), full_path: String::new() },
-                DirOrSong::Song(Song::default()),
-            ];
+    use super::browser::DirOrSong;
+    use crate::{
+        config::{
+            sort_mode::{SortMode, SortOptions},
+            theme::properties::SongProperty,
+        },
+        mpd::commands::{Song, metadata_tag::MetadataTag},
+    };
 
-            input.sort();
+    static LAST_ID: AtomicU32 = AtomicU32::new(1);
+    static NOW: LazyLock<chrono::DateTime<chrono::Utc>> = LazyLock::new(chrono::Utc::now);
+    pub fn new_id() -> u32 {
+        LAST_ID.fetch_add(1, std::sync::atomic::Ordering::Relaxed)
+    }
 
-            assert_eq!(input, vec![
-                DirOrSong::Dir { name: "a".to_owned(), full_path: String::new() },
-                DirOrSong::Dir { name: "z".to_owned(), full_path: String::new() },
-                DirOrSong::Song(Song::default()),
-                DirOrSong::Song(Song::default()),
-                DirOrSong::Song(Song::default()),
-            ]);
+    fn song(name: &str, metadata: &[(&str, &str)]) -> DirOrSong {
+        song_mtime(name, metadata, &NOW.to_rfc3339())
+    }
+
+    fn song_mtime(name: &str, metadata: &[(&str, &str)], mtime: &str) -> DirOrSong {
+        DirOrSong::Song(Song {
+            id: new_id(),
+            file: name.to_string(),
+            duration: Some(Duration::from_secs(1)),
+            metadata: metadata
+                .iter()
+                .map(|(k, v)| ((*k).to_string(), MetadataTag::Single((*v).to_string())))
+                .collect(),
+            stickers: None,
+            last_modified: mtime.parse().unwrap(),
+            added: None,
+        })
+    }
+
+    fn dir(name: &str) -> DirOrSong {
+        dir_mtime(name, &NOW.to_rfc3339())
+    }
+    fn dir_mtime(name: &str, mtime: &str) -> DirOrSong {
+        DirOrSong::Dir {
+            name: name.to_string(),
+            full_path: name.to_string(),
+            last_modified: mtime.parse().unwrap(),
         }
+    }
+
+    fn assert_equivalent(actual: &[DirOrSong], expected: &[DirOrSong]) {
+        assert_eq!(actual.len(), expected.len());
+        for (a, b) in actual.iter().zip(expected.iter()) {
+            assert_eq!(a.dir_name_or_file_name(), b.dir_name_or_file_name());
+        }
+    }
+
+    #[test]
+    fn group_dirs_first() {
+        let sort = SortOptions {
+            mode: SortMode::Format(vec![SongProperty::Title]),
+            group_directories_first: true,
+            reverse: false,
+        };
+        let input = [
+            song("song3", &[("title", "c")]),
+            dir("dir1"),
+            song("song2", &[("title", "b")]),
+            dir("dir3"),
+            song("song1", &[("title", "a")]),
+            dir("dir2"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            dir("dir1"),
+            dir("dir2"),
+            dir("dir3"),
+            song("song1", &[]),
+            song("song2", &[]),
+            song("song3", &[]),
+        ]);
+    }
+
+    #[test]
+    fn group_dirs_first_reversed() {
+        let sort = SortOptions {
+            mode: SortMode::Format(vec![SongProperty::Title]),
+            group_directories_first: true,
+            reverse: true,
+        };
+        let input = [
+            song("song3", &[("title", "c")]),
+            dir("dir1"),
+            song("song2", &[("title", "b")]),
+            dir("dir3"),
+            song("song1", &[("title", "a")]),
+            dir("dir2"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            song("song3", &[]),
+            song("song2", &[]),
+            song("song1", &[]),
+            dir("dir3"),
+            dir("dir2"),
+            dir("dir1"),
+        ]);
+    }
+
+    #[test]
+    fn no_grouping() {
+        let sort = SortOptions {
+            mode: SortMode::Format(vec![SongProperty::Title]),
+            group_directories_first: false,
+            reverse: false,
+        };
+        let input = [
+            song("song3", &[("title", "e")]),
+            song("song2", &[("title", "c")]),
+            dir("b_dir1"),
+            dir("f_dir3"),
+            song("song1", &[("title", "a")]),
+            dir("d_dir2"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            song("song1", &[]),
+            dir("b_dir1"),
+            song("song2", &[]),
+            dir("d_dir2"),
+            song("song3", &[]),
+            dir("f_dir3"),
+        ]);
+    }
+
+    #[test]
+    fn no_grouping_reversed() {
+        let sort = SortOptions {
+            mode: SortMode::Format(vec![SongProperty::Title]),
+            group_directories_first: false,
+            reverse: true,
+        };
+        let input = [
+            song("e", &[("title", "e")]),
+            song("c", &[("title", "c")]),
+            dir("b"),
+            dir("f"),
+            song("a", &[("title", "a")]),
+            dir("d"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            dir("f"),
+            song("e", &[]),
+            dir("d"),
+            song("c", &[]),
+            dir("b"),
+            song("a", &[]),
+        ]);
+    }
+
+    #[test]
+    fn group_dirs_mtime() {
+        let sort = SortOptions {
+            mode: SortMode::ModifiedTime,
+            group_directories_first: true,
+            reverse: false,
+        };
+        let input = [
+            song_mtime("e", &[], "2025-04-02T14:52:05Z"),
+            song_mtime("c", &[], "2025-04-02T14:52:03Z"),
+            dir_mtime("b", "2025-04-02T14:52:02Z"),
+            dir_mtime("f", "2025-04-02T14:52:06Z"),
+            song_mtime("a", &[], "2025-04-02T14:52:01Z"),
+            dir_mtime("d", "2025-04-02T14:52:04Z"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            dir("b"),
+            dir("d"),
+            dir("f"),
+            song("a", &[]),
+            song("c", &[]),
+            song("e", &[]),
+        ]);
+    }
+
+    #[test]
+    fn group_dirs_mtime_reversed() {
+        let sort = SortOptions {
+            mode: SortMode::ModifiedTime,
+            group_directories_first: true,
+            reverse: true,
+        };
+        let input = [
+            song_mtime("e", &[], "2025-04-02T14:52:05Z"),
+            song_mtime("c", &[], "2025-04-02T14:52:03Z"),
+            dir_mtime("b", "2025-04-02T14:52:02Z"),
+            dir_mtime("f", "2025-04-02T14:52:06Z"),
+            song_mtime("a", &[], "2025-04-02T14:52:01Z"),
+            dir_mtime("d", "2025-04-02T14:52:04Z"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            song("e", &[]),
+            song("c", &[]),
+            song("a", &[]),
+            dir("f"),
+            dir("d"),
+            dir("b"),
+        ]);
+    }
+
+    #[test]
+    fn no_grouping_dirs_mtime() {
+        let sort = SortOptions {
+            mode: SortMode::ModifiedTime,
+            group_directories_first: false,
+            reverse: false,
+        };
+        let input = [
+            song_mtime("e", &[], "2025-04-02T14:52:05Z"),
+            song_mtime("c", &[], "2025-04-02T14:52:03Z"),
+            dir_mtime("b", "2025-04-02T14:52:02Z"),
+            dir_mtime("f", "2025-04-02T14:52:06Z"),
+            song_mtime("a", &[], "2025-04-02T14:52:01Z"),
+            dir_mtime("d", "2025-04-02T14:52:04Z"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            song("a", &[]),
+            dir("b"),
+            song("c", &[]),
+            dir("d"),
+            song("e", &[]),
+            dir("f"),
+        ]);
+    }
+
+    #[test]
+    fn no_grouping_dirs_mtime_reversed() {
+        let sort = SortOptions {
+            mode: SortMode::ModifiedTime,
+            group_directories_first: false,
+            reverse: false,
+        };
+        let input = [
+            song_mtime("e", &[], "2025-04-02T14:52:05Z"),
+            song_mtime("c", &[], "2025-04-02T14:52:03Z"),
+            dir_mtime("b", "2025-04-02T14:52:02Z"),
+            dir_mtime("f", "2025-04-02T14:52:06Z"),
+            song_mtime("a", &[], "2025-04-02T14:52:01Z"),
+            dir_mtime("d", "2025-04-02T14:52:04Z"),
+        ];
+
+        let result = input
+            .into_iter()
+            .sorted_by(|a, b| a.with_custom_sort(&sort).cmp(&b.with_custom_sort(&sort)))
+            .collect_vec();
+
+        dbg!(&result);
+        assert_equivalent(&result, &[
+            song("a", &[]),
+            dir("b"),
+            song("c", &[]),
+            dir("d"),
+            song("e", &[]),
+            dir("f"),
+        ]);
     }
 }
 
@@ -1158,6 +1533,8 @@ mod format_tests {
                     ("artist".to_string(), "artist".into()),
                 ]),
                 stickers: None,
+                last_modified: chrono::Utc::now(),
+                added: None,
             };
 
             let result = format.as_string(Some(&song), "");
@@ -1188,6 +1565,8 @@ mod format_tests {
                     ("track".to_string(), "123".into()),
                 ]),
                 stickers: None,
+                last_modified: chrono::Utc::now(),
+                added: None,
             };
             let status = Status {
                 volume: Volume::new(123),