+++ src/ui/panes/playlists.rs
@@ -119,11 +119,8 @@ impl Pane for PlaylistsPane {
                         .list_playlists()
                         .context("Cannot list playlists")?
                         .into_iter()
-                        .map(|playlist| DirOrSong::Dir {
-                            name: playlist.name,
-                            full_path: String::new(),
-                        })
-                        .sorted()
+                        .sorted_by(|a, b| a.name.cmp(&b.name))
+                        .map(|playlist| DirOrSong::name_only(playlist.name))
                         .collect();
                     Ok(MpdQueryResult::DirOrSong { data: result, origin_path: None })
                 },
@@ -154,11 +151,8 @@ impl Pane for PlaylistsPane {
                                 .list_playlists()
                                 .context("Cannot list playlists")?
                                 .into_iter()
-                                .map(|playlist| DirOrSong::Dir {
-                                    name: playlist.name,
-                                    full_path: String::new(),
-                                })
-                                .sorted()
+                                .sorted_by(|a, b| a.name.cmp(&b.name))
+                                .map(|playlist| DirOrSong::name_only(playlist.name))
                                 .collect();
                             Ok(MpdQueryResult::DirOrSong { data: result, origin_path: None })
                         },